fix: correct assoc equality, reduce exported symbols, and improve rangemap printing
* assoc comparisons now honor an optional value-compare callback when checking equality between two assocs; falls back to pointer equality otherwise
* reduce exported/internal symbol surface (make map release and pointerset get helpers static) — **BREAKING** for code relying on those internal symbols
* rangemap diagnostics print pointer values as hex with proper uintptr_t casts for portability
* remove obsolete pointerqueue enumerator declaration

Files: src/assoc/mulle--assoc.c, src/map/mulle--map.c, src/mulle-container-callback.c, src/queue/mulle--pointerqueue.c, src/rangemap/mulle--rangemap.c, src/set/pointer/mulle--pointerset-generic.c. Dotfile and tooling changes staged were ignored for this message.

+++ test/array/array.c
@@ -482,7 +482,7 @@ static void   copy_array( void)
 static void   find( void)
 {
    struct mulle_array   array;
-   unsigned long        found;
+   uintptr_t            found;
    char                 *s1;
    char                 *s2;
    char                 *s3;
